+++ backend/eqpg/pg_test.go
@@ -1,4 +1,4 @@
-package pg
+package eqpg
 
 import (
 	"context"
